Fix functionality of Special:Sitematrix
DBlists were moved, but the accompanying entries had yet to be updated.
Callback for $wgConf did not take in account that the used approach
could fail when the dbname was not set explicit in
InitialiseSettings.php, so it now checks for dbname first, then the wiki
tags and finally the default.